requirements.txt
requests==2.24.0
PyJWT==1.7.1
rich==5.0.0